YAXArraysToolbox

Documentation for YAXArraysToolbox.

YAXArraysToolbox.aggregate_time

YAXArraysToolbox.altitude_mask_results_proc

YAXArraysToolbox.altitude_masking_proc

YAXArraysToolbox.masking_proc

YAXArraysToolbox.s4time

YAXArraysToolbox.space4time_proc

YAXArraysToolbox.aggregate_time — Method

Aggregate by time

Arguments:

cube_in YAXArray Cube.

time_axis: String. Name of the time axis.

new_resolution: String. New temporal resolution can be "day", "month", "year".

new_time_step: Int64. Time step to be computed in the new time series. e.g. new_resolution="day", new_time_step=8 will compute the function each 8 days. The new time dimension will only contain the days corresponding to the 8th day.

fun: String. Function to be applied to aggregate the time. It can be "median", "mean", "std", "var", "sum", "quant", "min", "max".

p: Float64 in the interval [0,1]. If fun=quant p is the value of the quantile.

skipMissing: Boolean. Skip missing values when aggregating the data. If all values are missing, NaN is returned.

skipnan: Boolean. Skip NaN values when aggregating the data. If all values are NaN, NaN is returned.

showprog: Boolean. Progress Bar.

max_cache: String. Maximum cache to read the data. It must be in MB e.g. "100MB" or in GB "10GB".

Examples

YAXArraysToolbox.altitude_mask_results_proc — Method

Topographical variability processor

Arguments:

cube_in_altitude : Altitude YAXARRAY with two variables mean, and sd.

lon_axis_name : String. Name of the longitude axis on the input cubes. By default lon_axis_name = "lon"

lat_axis_name : String. Name of the longitude axis on the input cubes. By default lon_axis_name = "lat"

variable_name : String. Name of the Variable containing the variables "mean", and "sd".

winsize: Edge size of the moving window on pixels. By default winsize = 5. E.g. winsize = 5 will produce a moving window with 5^2 pixels.

showprog: Show progress bar. By default showprog = true

Output:

The Topographical variability processor produces a YAXARRAY.cube with three Indicators:

v1: $v_{1}=\dfrac{1}{n}\sum_{i = 1}^{n}\sigma_{h,i}$

High values of $v_{1}$ indicate hilly terrain over the considered scale, which should be discarded from the analysis.

v2: $v_{2}=\vert \mu_{h} - \dfrac{1}{n} \sum_{i = 1}^{n}\mu_{h,i} \vert$

$v_{2}$ indicates how different the mean elevation within the central pixel is from the average elevation in the local window.

v3: $v_{3}= \vert \sigma_{h} - \dfrac{1}{n}\sum_{i = 1}^{n}\sigma_{h,i} \vert$

$v_{3}$: compares the central pixel's standard deviation of elevation with the standard deviation across the moving window.

See also:

altitude_masking_proc TO ADD LINK!!

Bibliography:

Duveiller, G., Hooker, J., & Cescatti, A. (2018). A dataset mapping the potential biophysical effects of vegetation cover change. Scientific Data, 5(1), Article 1. https://doi.org/10.1038/sdata.2018.14

YAXArraysToolbox.altitude_masking_proc — Method

Topographical masking processor

Arguments:

cube_in_to_mask: YAXArray Cube to be masked.

cube_in_altitude : Altitude YAXARRAY with two variables mean, and sd.

lon_axis_name : String. Name of the longitude axis on the input cubes. By default lon_axis_name = "lon"

lat_axis_name : String. Name of the longitude axis on the input cubes. By default lon_axis_name = "lat"

variable_name : String. Name of the Variable containing the variables "mean", and "sd".

time_axis_name : String or NaN. It is strongly recommended to pass this parameter if the cube to be masked contains a time dimension, otherwise nothing.

winsize: Edge size of the moving window on pixels. By default winsize = 5. E.g. winsize = 5 will produce a moving window with 5^2 pixels.

v1_thr : Float. Threshold to mask values using $v_{1}$ indicator. All values higer or equal to v1_thr are set to NaN. By default v1_thr = 50

v2_thr : Float. Threshold to mask values using $v_{2}$ indicator. All values higer or equal to v2_thr are set to NaN. By detault v2_thr = 100

v3_thr : Float. Threshold to mask values using $v_{3}$ indicator. All values higer or equal to v3 are set to NaN. By default v3_thr = 100

showprog: Boolean. Show progress bar. By default showprog = true.

Output:

YAXArray Datase with two variables:

cube masked: YAXArray Cube with same dimensions as cube_in_to_mask.

masked_pixels: YAXArray Cube with same lat, lon, dimensions as cube_masked but with a single boolean variable indicating if the pixel was masked or not.

Topographical variability indicators

v1: $v_{1}=\dfrac{1}{n}\sum_{i = 1}^{n}\sigma_{h,i}$

High values of $v_{1}$ indicate hilly terrain over the considered scale, which should be discarded from the analysis.

v2: $v_{2}=\vert \mu_{h} - \dfrac{1}{n} \sum_{i = 1}^{n}\mu_{h,i} \vert$

$v_{2}$ indicates how different the mean elevation within the central pixel is from the average elevation in the local window.

v3: $v_{3}= \vert \sigma_{h} - \dfrac{1}{n}\sum_{i = 1}^{n}\sigma_{h,i} \vert$

$v_{3}$: compares the central pixel's standard deviation of elevation with the standard deviation across the moving window.

See also

altitude_mask_results_proc function TO ADD LINK!!

Bibliography

Duveiller, G., Hooker, J., & Cescatti, A. (2018). A dataset mapping the potential biophysical effects of vegetation cover change. Scientific Data, 5(1), Article 1. https://doi.org/10.1038/sdata.2018.14

YAXArraysToolbox.masking_proc — Method

Masking processor

Arguments:

cube_in_to_mask: YAXArray cube to be masked.

cube_rsquare: Nothing, or YAXArray cube with the $R^{2}$ variable. If set to nothing no mask is applied

rsquare_thr: Float64. $R^{2}$ threshold. All values lower than rsquare_thr are set to NaN

cube_co_occurrence: Nothing, or YAXArray cube with the co-occurrence variable. If set to nothing no mask is applied.

co_occurence_thr: Float64. Co-occurence threshold. All values lower than co_occurence_thr are set to NaN

cube_delta: Nothing, or YAXArray cube with delta variable. If set to nothing no mask is applied.

minmax_delta: Tuple. Minimum and maximum thresholds of delta variable. Values lower and higher than the thresholds are set to NaN. It is also possible to set any of the thresholds as nothing e.g. (-1, nothing) or (nothing, 1) in these cases only one threshold is applied.

time_dim: Nothing, or String. Name of the time dimension. This dimensions needs to be present in all the cubes. If set to nothing no time dimension considered (It can result in slower computation time!). By default time_dim = time

showprog: Boolean. Show progress bar. By default showprog = true

Output:

YAXArray cube masked.

YAXArraysToolbox.s4time — Method

space4time(climatecube, pftscube, pft_list::Vector{String}, winsize = 5, minpxl = 100, minDiffPxlspercentage = 40)

Compute the space for time analysis for a given climate variable. ...

Arguments

climate_cube: YAXARRAY cube with dimenssions: lon, lat, time.
pfts_cube: YAXARRAY cube with dimenssions: pfts, lat,lon, time.
...

Output

Three output cubes are generated.
out1: Summary statistics. YAXARRAY cube where summary_stat axis contains: 
* "rsquare": XXXX 
* "cumulative_variance": XXXX
* "predicted": Prediction of Z for the central pixel with its real PFT combination.
out2:

#Examples

YAXArraysToolbox.space4time_proc — Method

Space for time processor

Arguments:

cube_con : YAXARRAY with the continous variable to be analyized.

cube_classes: YAXARRAY with the discrete classes to be used in the space4time.

time_axis_name : String or nothing. Name of the time axis on the input cubes. By default time_axis_name = "time". if time_axis_name = nothing, not time dimension considered.

lon_axis_name : String. Name of the longitude axis on the input cubes. By default lon_axis_name = "lon"

lat_axis_name : String. Name of the longitude axis on the input cubes. By default lon_axis_name = "lat"

classes_var_name : String. Name of the Variable containing the discrete classes. By default classes_var_name = "classes".

winsize: Edge size of the moving window on pixels. By default winsize = 5. E.g. winsize = 5 will produce a moving window with 5^2 pixels.

minpxl : Minimum number of pixels in the moving window. By default minpxl = 25. Change accordindly to your winsize parameter.

minDiffPxlspercentage: Percentage of minimum number pixels in the moving window that must have different compositions. Must be any value in the interval 30-100. By default minDiffPxlspercentage = 40

classes_vec: A string vector with the names of the classes on cube_classes to be used. e.g. from MPI-BGC internal structure classes_vec = ["Evergreen_Needleleaf_Forests", "Evergreen_Broadleaf_Forests", "Deciduous_Needleleaf_Forests", "Deciduous_Broadleaf_Forests", "Mixed_Forests", "Closed_Shrublands", "Open_Shrublands", "Woody_Savannas", "Savannas", "Grasslands", "Permanent_Wetlands", "Croplands", "Urban_and_Built-up_Lands", "Cropland/Natural_Vegetation_Mosaics", "Permanent_Snow_and_Ice", "Barren", "Water_Bodies"]

max_value: Indicates if the scale of the presence of the discrete classes if from 0 to 1 or 0 to 100 if max_value = 100 then the data is re-scaled from 0 to 1. By default max_value = 1

showprog: Show progress bar. By default showprog = true

max_cache: Size of the cache to allocate temporarily sections of the cubes. By default max_cache = 1e8

Output:

The space4time_proc produces a YAXARRAY.Dataset with three cubes:

SummaryStats cube has one axis summary_stat, and three variables:

rsquared:

cumulative_variance:

predicted:

metrics_for_classes cube has one axis Values of Z for pure classes, and two variables:

estimated:

estimated_error:

metricsfortransitions has two axis transitions (all the transitions by pairs between the different classes), and Differences with three variables:

delta:

delta_error:

coocurence: